Upgraded from Flask to Next.js
Flask was not easily scalable, so instead switched to Next.js completely discarding the option to use FastAPI and others. During this lost my commit history thus went through a tedious process of going to that branch creating a clone of that branch, deleting the main branch, renaming the cloned temp branch to main and pushing the prev comit again. Hopeful that thing will go correct this time :)